Object Weather and service refactor.
