+++ src/index.js
@@ -1,21 +1,7 @@
-class Weather {
-  constructor(){
-    this.appid = 'a21ef4017ff27a42723559dd93f1f28e';
-    this.endpoint = 'https://api.openweathermap.org/data/2.5/weather';
-  }
-  async getWeather(units, location){
-    let response = await fetch(`${this.endpoint}?q=${location}&units=${units}&APPID=${this.appid}`, {mode: 'cors'});
-    return response.json();
-  }
+import WeatherService from './js/weather-service';
+import Weather from './js/weather';
 
-  async getCelsius(location){
-    return await this.getWeather('Metric', location);
-  }
-
-  async getFahrenheit(location){
-    return await this.getWeather('Imperial', location);
-  }
-}
-
-app = new Weather();
-app.getCelsius("Guadalajara").then(res => console.log(res));
+//let weather = new WeatherService();
+//weather.getCelsius("Guadalajara").then(res => console.log(JSON.stringify(res)));
+let celsius = new Weather({"coord":{"lon":-103.33,"lat":20.67},"weather":[{"id":803,"main":"Clouds","description":"broken clouds","icon":"04d"}],"base":"stations","main":{"temp":18.04,"feels_like":16.48,"temp_min":17,"temp_max":20,"pressure":1022,"humidity":55},"visibility":9656,"wind":{"speed":1.87,"deg":209},"clouds":{"all":75},"dt":1580495146,"sys":{"type":1,"id":7128,"country":"MX","sunrise":1580477408,"sunset":1580517789},"timezone":-21600,"id":4005539,"name":"Guadalajara","cod":200});
+console.log(celsius);
